Move source to Maven
