--- plugin.yml
@@ -1,6 +1,0 @@
-name: Kaboom
-main: me.noodles.boom.MainBoom
-version: 0.2
-author: BGHDNetwork
-commands:
-   kaboom:
